refactor: Sample util coverage (#22997)
