+++ lib/util/sample.ts
